+++ projects/ngx-meta/src/twitter-card/index.ts
@@ -1,17 +1,3 @@
-// Main providers
-export * from './src/ngx-meta-twitter-card.module'
-export * from './src/provide-ngx-meta-twitter-card'
-// Types
-export * from './src/twitter-card'
-export * from './src/twitter-card-creator'
-export * from './src/twitter-card-image'
-export * from './src/twitter-card-metadata'
-export * from './src/twitter-card-site'
-export * from './src/twitter-card-type'
-// Specific metadata managers
-export * from './src/twitter-card-card-metadata-provider'
-export * from './src/twitter-card-site-metadata-provider'
-export * from './src/twitter-card-creator-metadata-provider'
-export * from './src/twitter-card-description-metadata-provider'
-export * from './src/twitter-card-title-metadata-provider'
-export * from './src/twitter-card-image-metadata-provider'
+export * from './src/providers'
+export * from './src/managers'
+export * from './src/types'
