+++ projects/ngx-meta/src/twitter-card/src/managers/index.ts
@@ -1,0 +1,17 @@
+export { TWITTER_CARD_CARD_METADATA_PROVIDER } from './twitter-card-card-metadata-provider'
+export * from './twitter-card-type'
+export { TWITTER_CARD_CREATOR_METADATA_PROVIDER } from './twitter-card-creator-metadata-provider'
+export {
+  TwitterCardCreator,
+  TwitterCardCreatorId,
+  TwitterCardCreatorUsername,
+} from './twitter-card-creator'
+export { TWITTER_CARD_DESCRIPTION_METADATA_PROVIDER } from './twitter-card-description-metadata-provider'
+export { TWITTER_CARD_IMAGE_METADATA_PROVIDER } from './twitter-card-image-metadata-provider'
+export { TwitterCardImage } from './twitter-card-image'
+export { TWITTER_CARD_SITE_METADATA_PROVIDER } from './twitter-card-site-metadata-provider'
+export {
+  TwitterCardSite,
+  TwitterCardSiteId,
+  TwitterCardSiteUsername,
+} from './twitter-card-site'
