+++ projects/ngx-meta/src/twitter-card/src/managers/twitter-card-creator.ts
