+++ projects/ngx-meta/src/twitter-card/src/managers/twitter-card-site-metadata-provider.ts
@@ -1,7 +1,7 @@
-import { makeTwitterCardMetadataProvider } from './make-twitter-card-metadata-provider'
-import { TwitterCard } from './twitter-card'
+import { makeTwitterCardMetadataProvider } from '../utils/make-twitter-card-metadata-provider'
+import { TwitterCard } from '../types/twitter-card'
 import { TwitterCardSiteId, TwitterCardSiteUsername } from './twitter-card-site'
-import { makeTwitterCardMetaDefinition } from './make-twitter-card-meta-definition'
+import { makeTwitterCardMetaDefinition } from '../utils/make-twitter-card-meta-definition'
 
 const KEY = 'site' satisfies keyof TwitterCard
