+++ projects/ngx-meta/src/twitter-card/src/providers/index.ts
@@ -1,0 +1,2 @@
+export { NgxMetaTwitterCardModule } from './ngx-meta-twitter-card.module'
+export { provideNgxMetaTwitterCard } from './provide-ngx-meta-twitter-card'
