+++ projects/ngx-meta/src/twitter-card/src/types/index.ts
@@ -1,0 +1,2 @@
+export { TwitterCard } from './twitter-card'
+export { TwitterCardMetadata } from './twitter-card-metadata'
