+++ projects/ngx-meta/src/twitter-card/src/utils/make-twitter-card-metadata-provider.ts
@@ -6,8 +6,8 @@ import {
   NgxMetaMetaService,
 } from '@davidlj95/ngx-meta/core'
 import { FactoryProvider } from '@angular/core'
-import { TwitterCard } from './twitter-card'
-import { TwitterCardMetadata } from './twitter-card-metadata'
+import { TwitterCard } from '../types/twitter-card'
+import { TwitterCardMetadata } from '../types/twitter-card-metadata'
 import { makeTwitterCardMetaDefinition } from './make-twitter-card-meta-definition'
 
 const TWITTER_KEY: keyof TwitterCardMetadata = `twitterCard`
